+++ lib/linux_pkg/linux_thumbnail.dart
@@ -2,15 +2,70 @@ import 'dart:io';
 import 'dart:math';
 
 import 'package:flutter/foundation.dart';
+import 'package:than_pkg/types/src_dist_type.dart';
 
 class LinuxThumbnail {
   static final LinuxThumbnail thumbnail = LinuxThumbnail._();
   LinuxThumbnail._();
   factory LinuxThumbnail() => thumbnail;
 
+  //new methods
+  Future<void> genVideoThumbnail({
+    required List<SrcDistType> pathList,
+    int iconSize = 300,
+    bool isOverride = false,
+  }) async {
+    try {
+      for (final video in pathList) {
+        final imageFile = File(video.dist);
+        if (isOverride == false) {
+          if (await imageFile.exists()) {
+            continue;
+          }
+        }
+        await Process.run('ffmpeg', [
+          '-i', video.src,
+          '-ss', '00:00:05', // Capture at 5 seconds
+          '-vframes', '1',
+          imageFile.path
+        ]);
+      }
+    } catch (e) {
+      debugPrint("genVideoCover: ${e.toString()}");
+    }
+  }
+
+  Future<void> genPdfThumbnail({
+    required List<SrcDistType> pathList,
+    int iconSize = 300,
+    bool isOverride = false,
+  }) async {
+    try {
+      for (final pdfPath in pathList) {
+        final imageFile = File(pdfPath.dist);
+
+        if (isOverride == false && await imageFile.exists()) {
+          continue;
+        }
+        await Process.run('pdftoppm', [
+          '-png',
+          '-f',
+          '1',
+          '-singlefile',
+          pdfPath.src,
+          imageFile.path.replaceAll('.png', '')
+        ]);
+      }
+    } catch (e) {
+      debugPrint("genPdfCover: ${e.toString()}");
+    }
+  }
+
+  // old methods
+
   Future<void> genVideoCoverList({
-    required Comparable<String> outDirPath,
     required List<String> videoPathList,
+    required String outDirPath,
     int iconSize = 300,
   }) async {
     try {