+++ lib/types/src_dist_type.dart
@@ -1,0 +1,13 @@
+class SrcDistType {
+  String src;
+  String dist;
+  SrcDistType({
+    required this.src,
+    required this.dist,
+  });
+
+  Map<String, dynamic> toMap() => {
+        'src': src,
+        'dist': dist,
+      };
+}